Store Grid dimensions as usize.

+++ src/legacy/terrain.rs
@@ -17,8 +17,8 @@ impl TerrainMap {
 }
 
 impl Grid2D<TerrainType> for TerrainMap {
-    const W: u16 = 1024;
-    const H: u16 = 1024;
+    const W: usize = 1024;
+    const H: usize = 1024;
 
     fn set(&mut self, _position: Coord, _value: TerrainType) {
         todo!()
